- SigninScreen 의 요소들을 uiState 로 관리 - SigninViewModel.kt 의 요소를 combine으로 합침 - SigninScreen 과 SigninViewModel 의 in 부분의 i를 소문자로 변경 - ChatChatScreen.kt 에 파일명 변경에 의한 코드 수정
